Add styling to the content box scroll
I added custom styles to the scrollbar of the content box to improve the visual appearance and user experience. The scrollbar now has a thin width and a blue thumb in WebKit browsers. In Firefox, it has a width of 8px and a rebeccapurple color for the bar and green color for the thumb. The thumb also changes to red when hovered over. This enhancement makes the content box look more polished and aesthetically pleasing.

Files modified:
- src/components/ContentBox.js